requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.3
fastapi==0.136.0
h11==0.16.0
idna==3.13
packaging==26.1
pillow==12.2.0
pydantic==2.13.3
pydantic_core==2.46.3
pytesseract==0.3.13
python-dotenv==1.0.1
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.45.0
opencv-python==4.10.0.84
python-multipart==0.0.20
